Fixed gitHash

+++ scripts/deploy-freightsapp.ps1
@@ -2,6 +2,10 @@ $location = Get-Location
 Set-Location -Path $PSScriptRoot/../src
 node ./grab-git-info.js
 pnpm run nx build freights-app --source-map=true --base-href=/pwa/
+printf "Build exited with code $LASTEXITCODE - $?"
+if ($?) {
+    return $LASTEXITCODE
+}
 rm -rf $PSScriptRoot/../websites/freights.app/pwa
 copy-item $PSScriptRoot/../src/dist/apps/freights-app $PSScriptRoot/../websites/freights.app/pwa -force -recurse
 Set-Location -Path $PSScriptRoot/..
